{"schema":"decodex/commit/1","summary":"Split quantitative product validation","authority":"manual"}
